parser_liste_csv/requirements.txt
pika==1.3.2
pandas==2.2.3
psycopg2-binary==2.9.9
